test of put

+++ controllers/users.js
@@ -20,14 +20,22 @@ module.exports.create = function(request, response, next) {
 
 // DELETE /user/:id
 module.exports.delete = function(request, response, next) {
+  if (request.session.user.id === request.params.id){
   User.findByIdAndDelete(request.params.id)
     .then(user => user ? response.status(200).end() : next())
     .catch(error => next(error));
+  } else {
+    response.status(403).send('Forbidden Access');
+  }
 };
 
 // PUT /user/:id (update team/add team to user)
 module.exports.update = function(request, response, next) {
+  if (request.session.user.id === request.params.id){
   User.findByIdAndUpdate(request.params.id, request.body)
     .then(course => course ? response.status(200).end() : next())
     .catch(error => next(error));
+  } else {
+    response.status(403).send('Forbidden Access');
+  }
 };
